backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = ""
authors = [
    {name = "tainanpenning",email = "tainanhartwigpenning@gmail.com"}
]
requires-python = ">=3.12,<4.0"
dependencies = [
    "django (>=6.0.4,<7.0.0)",
    "djangorestframework (>=3.17.1,<4.0.0)",
    "psycopg2-binary (>=2.9.12,<3.0.0)",
    "django-cors-headers (>=4.9.0,<5.0.0)",
    "pillow (>=12.2.0,<13.0.0)",
    "django-extensions (>=4.1,<5.0)"
]

[dependency-groups]
dev = [
    "pytest (>=9.0.3,<10.0.0)",
    "pytest-django (>=4.7.0,<5.0.0)",
    "factory-boy (>=3.3.3,<4.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
